add dataclass default config

+++ policies/mixed_precision.py
@@ -27,6 +27,6 @@ bfSixteen = MixedPrecision(
 
 bfSixteen_working = MixedPrecision(
     param_dtype=torch.float32,
-    reduce_dtype=torch.float32,
-    buffer_dtype=torch.float32,
+    reduce_dtype=torch.bfloat16,
+    buffer_dtype=torch.bfloat16,
 )
